Split color out of the shade header font update flag
These need to be two flags in case we decide not to ship shade blur.
The text color of the shade header should only be updated to a
dynamic value that responds to the theme if the blur flag is enabled,
because when it is disabled the shade header background is always
black with the font hard coded to white.

Test: manual
Bug: 405266952
Bug: 370555223
Flag: com.android.systemui.shade_header_blur_font_color
Change-Id: I698b875a174bfb62f97d0f02950846fb372997a0